rename folders for BulkEditView and SingleEditView
