+++ client/src/modules/lead/EditLead/EditLeadSidebar/EditCard/CardFields/EditView/BulkEditView/BulkEditView.css
